Update README.md
# Santet-dos

Santet Dos is Dos tools for Attack of one Service, but this tools for Auto Dos or Dos Pwn of vulnerability Services. A Denial-of-Service (DoS) attack is an attack meant to shut down a machine or network, making it inaccessible to its intended users. DoS attacks accomplish this by flooding the target with traffic, or sending it information that triggers a crash.

# Need :

- Nmap
- Metasploit
- Searchsploit/Exploit-DB

- Dos HTTP : HTTP vulnerability Service
- Dos Win : RDP vulnerability Service

- Run without Root